update link api and fix some bugs related

+++ lib/format/format.dart
@@ -3,13 +3,13 @@ import 'package:intl/intl.dart';
 class FormatProvider {
   String formatNumber(String input) {
     final formatter = NumberFormat("#,###");
-    final number = int.parse(input);
+    final number = double.parse(input);
     return formatter.format(number);
   }
 
   String formatPrice(String input) {
     final formatter = NumberFormat("#,###");
-    final number = int.parse(input);
+    final number = double.parse(input);
     return formatter.format(number);
   }
 }
